agent.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

def get_gemini_client():
    """
    Initializes and returns the OpenAI client.
    It looks for the OPENAI_API_KEY in the environment variables.
    You should have a .env file in the root of the project with:
    OPENAI_API_KEY="your_api_key"
    """
    load_dotenv()

    api_key = os.getenv("GEMINI_API_KEY")
    if not api_key:
        raise ValueError("GEMINI_API_KEY not found in environment variables. Please create a .env file with the key.")
    genai.configure(api_key=api_key)
    return genai.GenerativeModel(model_name="gemini-2.5-pro")

def initialize_mcp():
    """
    Placeholder for initializing the Context7 MCP.
    """
    # In a real implementation, you would initialize the Context7 MCP here.
    print("Context7 MCP initialization is not yet implemented.")

# You can add more agent-related functions here, for example, to manage
# conversations, memory, or to interact with other tools.
